# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDatabaseRequest": {
        "additionalProperties": false,
        "properties": {
          "database_id": {
            "description": "Workspace-relative identifier. Never a filesystem path.",
            "examples": [
              "demo"
            ],
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
            "title": "Database Id",
            "type": "string"
          },
          "page_size": {
            "default": 4096,
            "description": "Bytes per page; must be a power of two. Small values are useful for demos because they force page chaining after a few rows.",
            "maximum": 65536.0,
            "minimum": 256.0,
            "title": "Page Size",
            "type": "integer"
          }
        },
        "required": [
          "database_id"
        ],
        "title": "CreateDatabaseRequest",
        "type": "object"
      },
      "CreateTableRequest": {
        "additionalProperties": false,
        "description": "Milestone 1 has no SQL parser, so tables are defined structurally.\n\nMilestone 2 adds ``POST /query`` with ``CREATE TABLE``; this endpoint stays\nas the programmatic path.",
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnSpec"
            },
            "minItems": 1,
            "title": "Columns",
            "type": "array"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "columns"
        ],
        "title": "CreateTableRequest",
        "type": "object"
      },
      "DatabaseDetail": {
        "properties": {
          "database_id": {
            "title": "Database Id",
            "type": "string"
          },
          "file_size_bytes": {
            "title": "File Size Bytes",
            "type": "integer"
          },
          "format_version": {
            "title": "Format Version",
            "type": "integer"
          },
          "free_list_head": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free List Head"
          },
          "heap_page_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Heap Page Ids",
            "type": "array"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Live rows; null when no table is defined yet",
            "title": "Row Count"
          },
          "schema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchemaModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_page_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Schema Page Ids",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/PagerStatsModel"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          },
          "trace_level": {
            "title": "Trace Level",
            "type": "string"
          }
        },
        "required": [
          "database_id",
          "page_size",
          "page_count",
          "file_size_bytes",
          "format_version",
          "table_name",
          "row_count",
          "heap_page_ids",
          "schema_page_ids",
          "free_list_head",
          "stats",
          "trace_level"
        ],
        "title": "DatabaseDetail",
        "type": "object"
      },
      "DatabaseListResponse": {
        "properties": {
          "databases": {
            "items": {
              "$ref": "#/components/schemas/DatabaseSummary"
            },
            "title": "Databases",
            "type": "array"
          },
          "workspace": {
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "databases",
          "workspace"
        ],
        "title": "DatabaseListResponse",
        "type": "object"
      },
      "DatabaseSummary": {
        "properties": {
          "database_id": {
            "title": "Database Id",
            "type": "string"
          },
          "is_open": {
            "title": "Is Open",
            "type": "boolean"
          },
          "modified_ns": {
            "title": "Modified Ns",
            "type": "integer"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "database_id",
          "size_bytes",
          "modified_ns",
          "is_open"
        ],
        "title": "DatabaseSummary",
        "type": "object"
      },
      "DeleteRecordResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "record_id": {
            "$ref": "#/components/schemas/RecordIdModel"
          }
        },
        "required": [
          "deleted",
          "record_id"
        ],
        "title": "DeleteRecordResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InsertRecordsRequest": {
        "additionalProperties": false,
        "properties": {
          "rows": {
            "description": "Positional values per row, in column order",
            "items": {
              "items": {},
              "type": "array"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "InsertRecordsRequest",
        "type": "object"
      },
      "InsertRecordsResponse": {
        "properties": {
          "duration_ns": {
            "title": "Duration Ns",
            "type": "integer"
          },
          "inserted": {
            "title": "Inserted",
            "type": "integer"
          },
          "pages_allocated": {
            "title": "Pages Allocated",
            "type": "integer"
          },
          "record_ids": {
            "items": {
              "$ref": "#/components/schemas/RecordIdModel"
            },
            "title": "Record Ids",
            "type": "array"
          }
        },
        "required": [
          "inserted",
          "record_ids",
          "pages_allocated",
          "duration_ns"
        ],
        "title": "InsertRecordsResponse",
        "type": "object"
      },
...
      "PagerStatsModel": {
        "description": "Cumulative I/O counters since the database handle was opened.",
        "properties": {
          "allocations": {
            "title": "Allocations",
            "type": "integer"
          },
          "bytes_read": {
            "title": "Bytes Read",
            "type": "integer"
          },
          "bytes_written": {
            "title": "Bytes Written",
            "type": "integer"
          },
          "frees": {
            "title": "Frees",
            "type": "integer"
          },
          "page_reads": {
            "title": "Page Reads",
            "type": "integer"
          },
          "page_writes": {
            "title": "Page Writes",
            "type": "integer"
          },
          "read_time_ns": {
            "title": "Read Time Ns",
            "type": "integer"
          },
          "recycled_allocations": {
            "title": "Recycled Allocations",
            "type": "integer"
          },
          "syncs": {
            "title": "Syncs",
            "type": "integer"
          },
          "write_time_ns": {
            "title": "Write Time Ns",
            "type": "integer"
          }
        },
        "required": [
          "page_reads",
          "page_writes",
          "allocations",
          "recycled_allocations",
          "frees",
          "syncs",
          "bytes_read",
          "bytes_written",
          "read_time_ns",
          "write_time_ns"
        ],
        "title": "PagerStatsModel",
        "type": "object"
      },
...
      "RecordsResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnModel"
            },
            "title": "Columns",
            "type": "array"
          },
          "duration_ns": {
            "title": "Duration Ns",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "pages_read": {
            "description": "Page reads this request caused",
            "title": "Pages Read",
            "type": "integer"
          },
          "returned": {
            "title": "Returned",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RowModel"
            },
            "title": "Rows",
            "type": "array"
          },
          "rows_scanned": {
            "description": "Rows the heap scan touched, including those skipped by offset",
            "title": "Rows Scanned",
            "type": "integer"
          }
        },
        "required": [
          "columns",
          "rows",
          "offset",
          "limit",
          "returned",
          "has_more",
          "rows_scanned",
          "pages_read",
          "duration_ns"
        ],
        "title": "RecordsResponse",
        "type": "object"
      },
      "RowModel": {
        "properties": {
          "record_id": {
            "$ref": "#/components/schemas/RecordIdModel"
          },
          "values": {
            "description": "Positional values matching the table's column order; null is NULL",
            "items": {},
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "record_id",
          "values"
        ],
        "title": "RowModel",
        "type": "object"
      },
      "SchemaModel": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnModel"
            },
            "title": "Columns",
            "type": "array"
          },
          "fixed_row_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Constant encoded row size, or null if any column varies",
            "title": "Fixed Row Size"
          },
          "null_bitmap_size": {
            "description": "Bytes of null bitmap per record",
            "title": "Null Bitmap Size",
            "type": "integer"
          }
        },
        "required": [
          "columns",
          "null_bitmap_size",
          "fixed_row_size"
        ],
        "title": "SchemaModel",
        "type": "object"
      },
...
      "TableResponse": {
        "properties": {
          "first_page_id": {
            "title": "First Page Id",
            "type": "integer"
          },
          "heap_page_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Heap Page Ids",
            "type": "array"
          },
          "last_page_id": {
            "title": "Last Page Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "schema": {
            "$ref": "#/components/schemas/SchemaModel"
          }
        },
        "required": [
          "name",
          "schema",
          "row_count",
          "heap_page_ids",
          "first_page_id",
          "last_page_id"
        ],
        "title": "TableResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP and WebSocket access to a running ChenDB engine, for the Visual Database\nExplorer.\n\nEverything served here is real engine state read from a real database file.\nEndpoints for features that do not exist yet are absent rather than stubbed:\ncheck `features` on `/health` to see what this build supports.",
    "title": "ChenDB Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/databases": {
      "get": {
        "operationId": "list_databases_api_v1_databases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List databases",
        "tags": [
          "databases"
        ]
      },
      "post": {
        "operationId": "create_database_api_v1_databases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a database",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/databases/{database_id}": {
      "delete": {
        "operationId": "delete_database_api_v1_databases__database_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a database file",
        "tags": [
          "databases"
        ]
      },
      "get": {
        "operationId": "get_database_detail_api_v1_databases__database_id__get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Database detail",
        "tags": [
          "databases"
        ]
      }
    },
...
    "/api/v1/databases/{database_id}/records": {
      "get": {
        "description": "Return a window of the heap scan.\n\nThere is no index and no ``LIMIT`` pushdown until later milestones, so\n``offset`` is honoured by discarding rows the scan already produced. The\nresponse reports ``rows_scanned`` and ``pages_read`` precisely so that cost\nis visible rather than hidden.",
        "operationId": "list_records_api_v1_databases__database_id__records_get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan rows",
        "tags": [
          "databases"
        ]
      },
      "post": {
        "operationId": "insert_records_api_v1_databases__database_id__records_post",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertRecordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertRecordsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Insert rows",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/databases/{database_id}/records/{page_id}/{slot_id}": {
      "delete": {
        "operationId": "delete_record_api_v1_databases__database_id__records__page_id___slot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "title": "Slot Id",
              "type": "integer"
            }
          },
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete the row at a record id",
        "tags": [
          "databases"
        ]
      }
    },
    "/api/v1/databases/{database_id}/table": {
      "get": {
        "operationId": "get_table_api_v1_databases__database_id__table_get",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Table and schema",
        "tags": [
          "databases"
        ]
      },
      "post": {
        "operationId": "create_table_api_v1_databases__database_id__table_post",
        "parameters": [
          {
            "description": "Workspace-relative database id",
            "in": "path",
            "name": "database_id",
            "required": true,
            "schema": {
              "description": "Workspace-relative database id",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Database Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Define the table (Milestone 1 allows one per database)",
        "tags": [
          "databases"
        ]
      }
    },
...
  }
}